fast-api-backend/pyproject.toml
[project]
name = "fast-api-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.123.4",
    "uvicorn>=0.25.0",
    "pydantic>=2.7.2",
    "firebase-admin>=6.5.0",
    "requests>=2.32.4",
    "python-dotenv>=1.1.0",
]
